feat: Add a basic script to parse XML
Added the source XML data and the json output to the commit since
they are not very likely to change.

Changes to package json:
- @types/node : for editor autocompletion and other niceties.
- xml2js : simple library that parses XML data into a JSON
structure, necessary because Node.js does not have any XML parsing
in its stdlib. Since it is used to generate static data, it is fine
in devDependencies.